Correct (uint32_t*)↔(size_t*) prototype mismatch
In the ctypes binding for Index_CreateWithStream, change the last
parameter from ctypes.POINTER(ctypes.c_uint32) to
ctypes.POINTER(ctypes.c_size_t) for consistency with libspatialindex.

This fixes the test failure on 64-bit big-endian architectures that
prompted #220. Other inconsistencies noted in that issue will be fixed
in additional commits.

https://github.com/Toblerity/rtree/issues/220#issue-1145043392

+++ rtree/core.py
@@ -99,7 +99,7 @@ NEXTFUNC = ctypes.CFUNCTYPE(
     ctypes.POINTER(ctypes.POINTER(ctypes.c_double)),
     ctypes.POINTER(ctypes.c_uint32),
     ctypes.POINTER(ctypes.POINTER(ctypes.c_ubyte)),
-    ctypes.POINTER(ctypes.c_uint32),
+    ctypes.POINTER(ctypes.c_size_t),
 )
 
 rt.Index_CreateWithStream.argtypes = [ctypes.c_void_p, NEXTFUNC]
